renaming project repository (from plural) and inheriting from base repository class
